+++ sigils/__init__.py
@@ -1,5 +1,6 @@
 import re
 import threading
+from .tools import tools
 
 
 class Sigil:
@@ -39,47 +40,95 @@ class Sigil:
             self.pattern = re.compile(r'%\[(.*?)\]')
             self._cache.value = {template: self.pattern}
 
-    def _resolve(self, context, depth=0, execute=True, max_depth=6):
+    def _run_func_with_sigil_args(self, func, func_args, value, context, execute, max_depth, debug):
+        num_args = func.__code__.co_argcount
+        if func_args:
+            if debug:
+                print(f"Tool: {func}, value: {value}, func_args: {func_args}")
+            resolved_args = [Sigil(f'%[{arg}]').interpolate(
+                context, execute, max_depth
+            ) for arg in func_args]
+            if debug:
+                print(f"Resolved args: {resolved_args}")
+            if num_args > 0:
+                if resolved_args and '%[' not in resolved_args[0]:
+                    return func(resolved_args[0], *resolved_args[1:num_args - 1])
+                else:
+                    return func(func_args[0], *func_args[1:num_args - 1])
+            else:
+                return func()
+        else:
+            if num_args == 1:
+                return func(str(value))
+            else:
+                return func()
+
+    def _resolve(self, context, depth=0, execute=True, max_depth=6, debug=False):
         resolved = {}
         for match in self.pattern.findall(self.template):
+            if debug:
+                print(f"Found match: {match}")
             keys = match.split('.')
+            if debug:
+                print(f"Split keys: {keys}")
             value = context
-            function_args = []
-            try:
-                for i, key in enumerate(keys):
-                    if ':' in key:
-                        key, *function_args = key.split(':')
-                    literal = False
-                    if key.startswith('%'):
-                        key = key[1:]
-                        literal = True
-                    if literal:
-                        temp = key
-                    elif isinstance(value, dict):
-                        temp = value.get(key)
-                    elif isinstance(value, list) and key.isdigit():
-                        temp = value[int(key)]
+            func_args = []
+            for i, key in enumerate(keys):
+                if ':' in key:
+                    key_parts = key.split(':')
+                    key = key_parts[0]
+                    func_args = key_parts[1:]
+                if debug:
+                    print(f"Processing key: {key}, func_args: {func_args}")
+                literal = False
+                if key.startswith('%'):
+                    key = key[1:]
+                    literal = True
+                if literal:
+                    temp = key
+                elif value and isinstance(value, dict) and key in value:
+                    if debug:
+                        print(f"Value is a dict with the key: {value}")
+                    temp = value.get(key, None)
+                    if callable(temp):
+                        temp = self._run_func_with_sigil_args(
+                            temp, func_args, value, context, execute, max_depth, debug)
+                elif value and isinstance(value, list) and key.isdigit():
+                    if debug:
+                        print(f"Value is a list: {value}")
+                    temp = value[int(key)]
+                elif key in tools:
+                    if debug:
+                        print(f"Found tool: {key}")
+                    tool_func = tools[key]
+                    if callable(tool_func):
+                        temp = self._run_func_with_sigil_args(
+                            tool_func, func_args, value, context, execute, max_depth, debug)
                     else:
-                        temp = None
-                    if temp is None and '-' in key and not literal:
-                        temp = value.get(key.replace('-', '_')) if isinstance(value, dict) else None
-                    if temp is None and hasattr(value, key) and not literal:
-                        temp = getattr(value, key)
-                    if temp is None and '-' in key and hasattr(value, key.replace('-', '_')) and not literal:
-                        temp = getattr(value, key.replace('-', '_'))
-                    if temp is not None:
-                        value = temp
-                        if callable(value) and execute:
-                            if function_args:
-                                resolved_function_args = [Sigil(f'%[{arg}]').interpolate(context) for arg in function_args]
-                                value = value(*resolved_function_args)
-                            else:
-                                value = value()
-                    else:
-                        value = None
-                        break
-            except (TypeError, AttributeError):
-                value = None
+                        if debug:
+                            print(f"Non-callable tool: {tool_func}")
+                        temp = tool_func
+                else:
+                    if debug:
+                        print(f"Value is an object: {value}")
+                    temp = None
+                if debug:
+                    print(f"Temp value: {temp}")
+                if temp and callable(temp):
+                    temp = temp()
+                if temp is None and '-' in key and not literal:
+                    temp = value.get(key.replace('-', '_')) if isinstance(value, dict) else None
+                if temp is None and hasattr(value, key) and not literal:
+                    temp = getattr(value, key)
+                if temp is None and '-' in key and hasattr(value, key.replace('-', '_')) and not literal:
+                    temp = getattr(value, key.replace('-', '_'))
+                if temp is not None:
+                    value = temp
+                else:
+                    if debug:
+                        print(f"Could not find value for key: {key}")
+                    value = None
+                    break
             if value is not None:
                 resolved[match] = value
             else:
@@ -90,34 +139,20 @@ class Sigil:
         if depth < max_depth:
             for key, value in list(resolved.items()):
                 if isinstance(value, str) and '%' in value:
+                    if debug:
+                        print(f"Found nested sigil: {value}")
                     s = Sigil(value)
-                    resolved_value = s._resolve(context, depth + 1, execute, max_depth)
+                    resolved_value = s._resolve(context, depth + 1, execute, max_depth, debug)
                     if resolved_value:
                         resolved[key] = {
-                            'value': s.interpolate(context, execute, max_depth),
+                            'value': s.interpolate(context, execute, max_depth, debug),
                             'sub_values': resolved_value
                         }
                     else:
                         resolved[key] = {'value': value}
         return resolved
 
-    def _get_value(self, key, value):
-        if isinstance(value, dict):
-            temp = value.get(key)
-            if temp is None and '-' in key:
-                temp = value.get(key.replace('-', '_'))
-        elif isinstance(value, list):
-            try:
-                temp = value[int(key)]
-            except (IndexError, ValueError):
-                temp = None
-        else:
-            temp = getattr(value, key, None)
-            if temp is None and '-' in key:
-                temp = getattr(value, key.replace('-', '_'), None)
-        return temp
-
-    def interpolate(self, context, execute=True, max_depth=6, handle_errors="propagate"):
+    def interpolate(self, context, execute=True, max_depth=6, handle_errors="propagate", debug=False):
         """
         Interpolate the template with the provided context.
 
@@ -127,7 +162,7 @@ class Sigil:
             max_depth (int, optional): The maximum depth for resolving nested sigils. Defaults to 6.
             handle_errors (str, optional): How to handle errors that occur during interpolation. 
                 Can be "propagate" (raise the error), "ignore" (return the original template), 
-                or "replace" (replace the template with the error message). Defaults to "propagate".
+                "replace" (replace the template with the error message). Defaults to "propagate".
 
         Returns:
             str: The interpolated string.
@@ -147,7 +182,7 @@ class Sigil:
         if context is None:
             context = Sigil.Context.current_context
         try:
-            resolved = self._resolve(context, 0, execute, max_depth)
+            resolved = self._resolve(context, 0, execute, max_depth, debug)
         except Exception as e:
             if handle_errors == "propagate":
                 raise e
@@ -206,9 +241,7 @@ class Sigil:
         ```
 
         """
-        print(f"Context: {context}")
-        result = self.interpolate(context)
-        print(f"Interpolated result: {result}") 
+        result = self.interpolate(context or {})
         return result